# Improve CLI startup time